+++ Changelog.md
@@ -2,6 +2,7 @@
 -------------
 - Bugfix: Intersections of countries and subgrids sometimes took an additional row and column causing a small fraction of the emissions to end up outside of the grid
 - Calculated PMCoarse value will match the PM10 value if no PM2.5 data is available
+- Sherpa EMEP grid added
 
 Release 3.2.1
 -------------